backend/requirements.txt
Django==5.2.5 
djangorestframework==3.16.1 
drf-yasg==1.21.7
python-dotenv==1.0.1
